requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
statsmodels==0.14.1
requests==2.31.0
python-dotenv==1.0.0
xlsxwriter==3.1.9
scikit-learn==1.3.2
pytz==2023.3
sendgrid==6.10.0
python-telegram-bot==20.7
nest_asyncio==1.5.8